+++ src/Connection/FormsConnection.php
@@ -53,8 +53,9 @@ class FormsConnection extends AbstractConnection {
 		return [
 			'formIds' => [
 				'type'        => [ 'list_of' => 'ID' ],
-				'description' => __( 'Array of form IDs to return. Exclude this argument to query all forms.', 'wp-graphql-gravity-forms' ),
+				'description' => __( 'Array of form database IDs to return. Exclude this argument to query all forms.', 'wp-graphql-gravity-forms' ),
 			],
+			// @todo make filtering more intuitive.
 			'status'  => [
 				'type'        => FormStatusEnum::$type,
 				'description' => __( 'Status of the forms to get.', 'wp-graphql-gravity-forms' ),